Crate varint

An implementation of Google Protobuf's Variable-Length Integers

Constants

VARINT_32_MAX_BYTES
The maximum number of bytes used by a 32-bit Varint

VARINT_64_MAX_BYTES
The maximum number of bytes used by a 32-bit Varint

Traits

VarintRead
Extends the I/O Read trait to provide functions to read (currently only 32-bit) Variable-Length Integers

VarintReader
Extends the IoOperations Reader trait to provide functions to read (currently only 32-bit) Variable-Length Integers

VarintWrite
Extends the I/O Write trait to provide functions for writing (currently only 32-bit) variable-length integers

VarintWriter
Extends the IoOperations Writer to provide functions for writing (currently only 32-bit) variable-length integers

ZigZag
A trait for enabling zig-zag encoding of various values